com - 1.6.8

+++ web-app/rest-service/src/main/java/karmanchik/chtotib/restservice/rest/file/FileImportController.java
@@ -87,7 +87,7 @@ public class FileImportController {
                                             .findFirst()
                                             .orElseThrow(() -> new ResourceNotFoundException(s, Teacher.class)));
                                 } else {
-                                    throw new StringReadException(s, "Иванов А.А.");
+                                    EXCEPTION_LIST.add(new StringReadException(s, "Иванов И.И.").getMessage());
                                 }
                             }
                         }
@@ -100,13 +100,24 @@ public class FileImportController {
                                 .pairNumber(pair)
                                 .build());
                     });
-            replacementRepository.deleteAll();
-            log.info("Save replacements [{}]...", replacements.size());
-            return ResponseEntity.ok()
-                    .body(replacementRepository.saveAll(replacements));
+            if (EXCEPTION_LIST.isEmpty()) {
+                replacementRepository.deleteAll();
+                log.info("Save replacements [{}]...", replacements.size());
+                replacementRepository.saveAll(replacements);
+                return ResponseEntity.ok()
+                        .body("OK");
+            }
+            else {
+                return ResponseEntity.badRequest()
+                        .body(Map.of(
+                                "status", "FAIL",
+                                "trace", EXCEPTION_LIST
+                        ));
+            }
         } catch (Exception e) {
             log.error(e.getMessage(), e);
-            return ResponseEntity.badRequest().body(e.getMessage());
+            return ResponseEntity.badRequest()
+                    .body(e.getMessage());
         }
     }
